+++ backend/gonlink/src/main/java/online/gonlink/entity/RealTimeTraffic.java
@@ -14,11 +14,11 @@ public class RealTimeTraffic implements Traffic{
     @Id
     private String shortCode;
     private String updateAt;
-    private int[] trafficMinute;
+    private short[] trafficMinute;
 
     public RealTimeTraffic(String shortCode, String updateAt) {
         this.shortCode = shortCode;
-        this.trafficMinute = new int[60];
+        this.trafficMinute = new short[60];
         this.updateAt = updateAt;
     }
 }
